requirements.txt
python-dotenv
flask
psycopg2-binary
pandas